Require users to be logged in/admins for queries/mutations

+++ src/graphql/resolvers/user/deleteUserResolver.js
@@ -1,13 +1,9 @@
 const models = require("../../../models");
-const Sequelize = require('sequelize');
 
-module.exports = (source, {id}, {tokenPayload}) => {
-
-  //Sequelize.query("UPDATE SQLITE_SEQUENCE SET SEQ=0 WHERE NAME='Users'");
-
-  // if(!tokenPayload) { // || !(tokenPayload.role === 'ADMIN')
-  //   return null;
-  // }
+module.exports = (source, {id}, ctx) => {
+  if (!ctx.tokenPayload || ctx.tokenPayload.role !== "ADMIN") {
+    throw new Error("Not allowed");
+  }
 
   return models.User.destroy({
     where: {
